refactor[forwarder](http): move the listeners package under collector
The HTTP/HTTPS collector lived in a top-level listeners package while the
syslog, netflow and file collectors lived under collector/. It implements the
same Collector interface and already depended on collector/configwatcher and
collector/schema, so it belongs beside its siblings.

The package name also claimed more than it held: it contained one collector,
while syslog and netflow have listeners too.

No behavior change. Only the package clause and the two import sites move; the
redundant import alias in collector.go and config.go is gone as a result.
