+++ collectors/forwarder/collector/http/auth.go
@@ -1,4 +1,4 @@
-package listeners
+package http
 
 import (
 	"crypto/hmac"
