+++ collectors/forwarder/collector/http/handler.go
@@ -1,4 +1,4 @@
-package listeners
+package http
 
 import (
 	"io"
