+++ collectors/forwarder/collector/http/http.go
@@ -1,4 +1,4 @@
-package listeners
+package http
 
 import (
 	"context"
